test that primary source filters are resolved to titles
