+++ tests/search/test_main.py
@@ -1,14 +1,19 @@
-import re
-
 import pytest
 from playwright.sync_api import Page, expect
 
-from mex.common.models import AnyExtractedModel
+from mex.common.models import (
+    MEX_PRIMARY_SOURCE_STABLE_TARGET_ID,
+    AnyExtractedModel,
+    ExtractedActivity,
+    ExtractedPrimarySource,
+)
 
 
 @pytest.mark.integration
 @pytest.mark.usefixtures("load_dummy_data")
-def test_index(frontend_url: str, writer_user_page: Page) -> None:
+def test_index(
+    frontend_url: str, writer_user_page: Page, extracted_activity: ExtractedActivity
+) -> None:
     page = writer_user_page
 
     # load page and establish section is visible
@@ -21,13 +26,13 @@ def test_index(frontend_url: str, writer_user_page: Page) -> None:
     expect(page.get_by_text("Showing 7 of 7 items")).to_be_visible()
 
     # check mex primary source is showing
-    primary_source = page.get_by_text(re.compile(r"^PrimarySource$"))
+    primary_source = page.get_by_test_id(
+        f"result-{MEX_PRIMARY_SOURCE_STABLE_TARGET_ID}"
+    )
     expect(primary_source.first).to_be_visible()
 
     # check activity is showing
-    activity = page.get_by_text(
-        re.compile(r"Aktivität 1\s*de\s*A1.*1999-12-24\s*day\s*2023-01-01")
-    )
+    activity = page.get_by_test_id(f"result-{extracted_activity.stableTargetId}")
     activity.scroll_into_view_if_needed()
     expect(activity).to_be_visible()
     page.screenshot(path="tests_search_test_main-test_index-focus-activity.png")
@@ -112,25 +117,35 @@ def test_entity_types(
 def test_had_primary_sources(
     frontend_url: str,
     writer_user_page: Page,
+    dummy_data_by_identifier_in_primary_source: dict[str, AnyExtractedModel],
 ) -> None:
     page = writer_user_page
     page.goto(frontend_url)
+
+    extracted_primary_source_one = dummy_data_by_identifier_in_primary_source["ps-1"]
+    assert isinstance(extracted_primary_source_one, ExtractedPrimarySource)
 
     # check sidebar is showing
     sidebar = page.get_by_test_id("search-sidebar")
     expect(sidebar).to_be_visible()
 
-    # check entity types are showing and functioning
+    # check primary sources are showing and functioning
     primary_sources = page.get_by_test_id("had-primary-sources")
+    primary_sources.scroll_into_view_if_needed()
     expect(primary_sources).to_be_visible()
     assert "00000000000000" in primary_sources.all_text_contents()[0]
+    # check that title is resolved if primary source has a title
+    assert (
+        extracted_primary_source_one.title[0].value
+        in primary_sources.all_text_contents()[0]
+    )
 
-    primary_sources.get_by_text("00000000000000").click()
+    primary_sources.get_by_text("Primary Source One").click()
     expect(page.get_by_text("Showing 3 of 3 items")).to_be_visible()
     page.screenshot(
         path="tests_search_test_main-test_had_primary_sources-on-select-primary-source-1-found.png"
     )
-    primary_sources.get_by_text("00000000000000").click()
+    primary_sources.get_by_text("Primary Source One").click()
 
 
 @pytest.mark.integration
